+++ tests/Geometry/Trigonometry/ArcSineTest.php
@@ -4,9 +4,9 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Math\Geometry\Trigonometry;
 
 use Innmind\Math\{
-    Geometry\Trigonometry\ArcSine,
-    Geometry\Trigonometry\Sine,
+    Geometry\Trigonometry,
     Geometry\Angle\Degree,
+    Geometry\Angle\Radian,
     Algebra\Number,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
@@ -15,9 +15,9 @@ class ArcSineTest extends TestCase
 {
     public function testInterface()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        );
+        $asin = Degree::of(Number::of(42))
+            ->sine()
+            ->apply(Trigonometry::arcSine);
 
         $this->assertInstanceOf(Degree::class, $asin->toDegree());
         $this->assertSame('sin⁻¹(sin(42°))', $asin->toString());
@@ -26,9 +26,13 @@ class ArcSineTest extends TestCase
 
     public function testEquals()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
 
         $this->assertTrue($asin->equals(Number::of(42.0)));
         $this->assertFalse($asin->equals(Number::of(0.74)));
@@ -36,9 +40,13 @@ class ArcSineTest extends TestCase
 
     public function testHigherThan()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
 
         $this->assertFalse($asin->higherThan(Number::of(42.0)));
         $this->assertTrue($asin->higherThan(Number::of(0.74)));
@@ -46,9 +54,13 @@ class ArcSineTest extends TestCase
 
     public function testAdd()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->add(Number::of(1));
 
         $this->assertSame(43.0, $number->value());
@@ -56,9 +68,13 @@ class ArcSineTest extends TestCase
 
     public function testSubtract()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->subtract(Number::of(66));
 
         $this->assertSame(-24.0, $number->value());
@@ -66,9 +82,13 @@ class ArcSineTest extends TestCase
 
     public function testDivideBy()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->divideBy(Number::of(2));
 
         $this->assertSame(21.0, $number->value());
@@ -76,9 +96,13 @@ class ArcSineTest extends TestCase
 
     public function testMulitplyBy()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->multiplyBy(Number::of(2));
 
         $this->assertSame(84.0, $number->value());
@@ -86,9 +110,13 @@ class ArcSineTest extends TestCase
 
     public function testRound()
     {
-        $number = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $number = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
 
         $this->assertEquals(42.0, $number->roundUp(1)->value());
         $this->assertEquals(42.0, $number->roundDown(1)->value());
@@ -98,9 +126,13 @@ class ArcSineTest extends TestCase
 
     public function testFloor()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->floor();
 
         $this->assertSame(42.0, $number->value());
@@ -108,9 +140,13 @@ class ArcSineTest extends TestCase
 
     public function testCeil()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->ceil();
 
         $this->assertSame(42.0, $number->value());
@@ -118,9 +154,13 @@ class ArcSineTest extends TestCase
 
     public function testModulo()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->modulo(Number::of(3));
 
         $this->assertSame(0.0, $number->value());
@@ -128,9 +168,13 @@ class ArcSineTest extends TestCase
 
     public function testAbsolute()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->absolute();
 
         $this->assertSame(42.0, $number->value());
@@ -138,9 +182,13 @@ class ArcSineTest extends TestCase
 
     public function testPower()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->power(Number::of(2));
 
         $this->assertSame(1764.0, $number->value());
@@ -148,9 +196,13 @@ class ArcSineTest extends TestCase
 
     public function testSquareRoot()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->squareRoot();
 
         $this->assertSame(6.48074069840786, $number->value());
@@ -158,9 +210,13 @@ class ArcSineTest extends TestCase
 
     public function testExponential()
     {
-        $asin = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
-        )->number();
+        $asin = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
+        )
+            ->toDegree()
+            ->number();
         $number = $asin->exponential();
 
         $this->assertSame(1.739274941520501E+18, $number->value());
@@ -168,9 +224,12 @@ class ArcSineTest extends TestCase
 
     public function testBinaryLogarithm()
     {
-        $number = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
+        $number = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
         )
+            ->toDegree()
             ->number()
             ->binaryLogarithm();
 
@@ -179,9 +238,12 @@ class ArcSineTest extends TestCase
 
     public function testNaturalLogarithm()
     {
-        $number = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
+        $number = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
         )
+            ->toDegree()
             ->number()
             ->naturalLogarithm();
 
@@ -190,9 +252,12 @@ class ArcSineTest extends TestCase
 
     public function testCommonLogarithm()
     {
-        $number = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
+        $number = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
         )
+            ->toDegree()
             ->number()
             ->commonLogarithm();
 
@@ -201,9 +266,12 @@ class ArcSineTest extends TestCase
 
     public function testSignum()
     {
-        $number = ArcSine::of(
-            Sine::of(Degree::of(Number::of(42)))->number(),
+        $number = Radian::of(
+            Degree::of(Number::of(42))
+                ->sine()
+                ->apply(Trigonometry::arcSine),
         )
+            ->toDegree()
             ->number()
             ->signum();