+++ tests/Geometry/Trigonometry/CosineTest.php
@@ -4,7 +4,6 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Math\Geometry\Trigonometry;
 
 use Innmind\Math\{
-    Geometry\Trigonometry\Cosine,
     Geometry\Angle\Degree,
     Algebra\Number,
 };
@@ -14,15 +13,15 @@ class CosineTest extends TestCase
 {
     public function testInterface()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)));
+        $cos = Degree::of(Number::of(42))->cosine();
 
-        $this->assertSame(0.7431448254773942, $cos->number()->value());
+        $this->assertSame(0.7431448254773942, $cos->value());
         $this->assertSame('cos(42°)', $cos->toString());
     }
 
     public function testEquals()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
 
         $this->assertTrue($cos->equals($cos));
         $this->assertTrue($cos->equals(Number::of(0.7431448254773942)));
@@ -31,7 +30,7 @@ class CosineTest extends TestCase
 
     public function testHigherThan()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
 
         $this->assertFalse($cos->higherThan($cos));
         $this->assertFalse($cos->higherThan(Number::of(0.7431448254773942)));
@@ -40,7 +39,7 @@ class CosineTest extends TestCase
 
     public function testAdd()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->add(Number::of(1));
 
         $this->assertSame(1.7431448254773942, $number->value());
@@ -48,7 +47,7 @@ class CosineTest extends TestCase
 
     public function testSubtract()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->subtract(Number::of(66));
 
         $this->assertSame(-65.2568551745226, $number->value());
@@ -56,7 +55,7 @@ class CosineTest extends TestCase
 
     public function testDivideBy()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->divideBy(Number::of(2));
 
         $this->assertSame(0.3715724127386971, $number->value());
@@ -64,7 +63,7 @@ class CosineTest extends TestCase
 
     public function testMulitplyBy()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->multiplyBy(Number::of(2));
 
         $this->assertSame(1.4862896509547885, $number->value());
@@ -72,7 +71,7 @@ class CosineTest extends TestCase
 
     public function testRound()
     {
-        $number = Cosine::of(Degree::of(Number::of(42)))->number();
+        $number = Degree::of(Number::of(42))->cosine();
 
         $this->assertEquals(0.7, $number->roundUp(1)->value());
         $this->assertEquals(0.7, $number->roundDown(1)->value());
@@ -82,7 +81,7 @@ class CosineTest extends TestCase
 
     public function testFloor()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->floor();
 
         $this->assertSame(0.0, $number->value());
@@ -90,7 +89,7 @@ class CosineTest extends TestCase
 
     public function testCeil()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->ceil();
 
         $this->assertSame(1.0, $number->value());
@@ -98,7 +97,7 @@ class CosineTest extends TestCase
 
     public function testModulo()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->modulo(Number::of(3));
 
         $this->assertSame(0.7431448254773942, $number->value());
@@ -106,7 +105,7 @@ class CosineTest extends TestCase
 
     public function testAbsolute()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->absolute();
 
         $this->assertSame(0.7431448254773942, $number->value());
@@ -114,7 +113,7 @@ class CosineTest extends TestCase
 
     public function testPower()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->power(Number::of(2));
 
         $this->assertSame(0.5522642316338268, $number->value());
@@ -122,7 +121,7 @@ class CosineTest extends TestCase
 
     public function testSquareRoot()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->squareRoot();
 
         $this->assertSame(0.8620584814717586, $number->value());
@@ -130,7 +129,7 @@ class CosineTest extends TestCase
 
     public function testExponential()
     {
-        $cos = Cosine::of(Degree::of(Number::of(42)))->number();
+        $cos = Degree::of(Number::of(42))->cosine();
         $number = $cos->exponential();
 
         $this->assertSame(2.1025372410974477, $number->value());
@@ -138,8 +137,8 @@ class CosineTest extends TestCase
 
     public function testBinaryLogarithm()
     {
-        $number = Cosine::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->cosine()
             ->binaryLogarithm();
 
         $this->assertSame(-0.42828470156966353, $number->value());
@@ -147,8 +146,8 @@ class CosineTest extends TestCase
 
     public function testNaturalLogarithm()
     {
-        $number = Cosine::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->cosine()
             ->naturalLogarithm();
 
         $this->assertSame(-0.29686433336996987, $number->value());
@@ -156,8 +155,8 @@ class CosineTest extends TestCase
 
     public function testCommonLogarithm()
     {
-        $number = Cosine::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->cosine()
             ->commonLogarithm();
 
         $this->assertSame(-0.1289265418564653, $number->value());
@@ -165,8 +164,8 @@ class CosineTest extends TestCase
 
     public function testSignum()
     {
-        $number = Cosine::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->cosine()
             ->signum();
 
         $this->assertSame(1, $number->value());