+++ tests/Geometry/Trigonometry/SineTest.php
@@ -4,7 +4,6 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Math\Geometry\Trigonometry;
 
 use Innmind\Math\{
-    Geometry\Trigonometry\Sine,
     Geometry\Angle\Degree,
     Algebra\Number,
 };
@@ -14,7 +13,7 @@ class SineTest extends TestCase
 {
     public function testInterface()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)));
+        $sin = Degree::of(Number::of(42))->sine();
 
         $this->assertSame(0.6691306063588582, $sin->number()->value());
         $this->assertSame('sin(42°)', $sin->toString());
@@ -22,7 +21,7 @@ class SineTest extends TestCase
 
     public function testEquals()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
 
         $this->assertTrue($sin->equals($sin));
         $this->assertTrue($sin->equals(Number::of(0.6691306063588582)));
@@ -31,7 +30,7 @@ class SineTest extends TestCase
 
     public function testHigherThan()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
 
         $this->assertFalse($sin->higherThan($sin));
         $this->assertFalse($sin->higherThan(Number::of(0.6691306063588582)));
@@ -40,7 +39,7 @@ class SineTest extends TestCase
 
     public function testAdd()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->add(Number::of(1));
 
         $this->assertSame(1.6691306063588582, $number->value());
@@ -48,7 +47,7 @@ class SineTest extends TestCase
 
     public function testSubtract()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->subtract(Number::of(66));
 
         $this->assertSame(-65.33086939364114, $number->value());
@@ -56,7 +55,7 @@ class SineTest extends TestCase
 
     public function testDivideBy()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->divideBy(Number::of(2));
 
         $this->assertSame(0.3345653031794291, $number->value());
@@ -64,7 +63,7 @@ class SineTest extends TestCase
 
     public function testMulitplyBy()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->multiplyBy(Number::of(2));
 
         $this->assertSame(1.3382612127177165, $number->value());
@@ -72,7 +71,7 @@ class SineTest extends TestCase
 
     public function testRound()
     {
-        $number = Sine::of(Degree::of(Number::of(42)))->number();
+        $number = Degree::of(Number::of(42))->sine();
 
         $this->assertEquals(0.7, $number->roundUp(1)->value());
         $this->assertEquals(0.7, $number->roundDown(1)->value());
@@ -82,7 +81,7 @@ class SineTest extends TestCase
 
     public function testFloor()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->floor();
 
         $this->assertSame(0.0, $number->value());
@@ -90,7 +89,7 @@ class SineTest extends TestCase
 
     public function testCeil()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->ceil();
 
         $this->assertSame(1.0, $number->value());
@@ -98,7 +97,7 @@ class SineTest extends TestCase
 
     public function testModulo()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->modulo(Number::of(3));
 
         $this->assertSame(0.6691306063588582, $number->value());
@@ -106,7 +105,7 @@ class SineTest extends TestCase
 
     public function testAbsolute()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->absolute();
 
         $this->assertSame(0.6691306063588582, $number->value());
@@ -114,7 +113,7 @@ class SineTest extends TestCase
 
     public function testPower()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->power(Number::of(2));
 
         $this->assertSame(0.4477357683661733, $number->value());
@@ -122,7 +121,7 @@ class SineTest extends TestCase
 
     public function testSquareRoot()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->squareRoot();
 
         $this->assertSame(0.8180040381066943, $number->value());
@@ -130,7 +129,7 @@ class SineTest extends TestCase
 
     public function testExponential()
     {
-        $sin = Sine::of(Degree::of(Number::of(42)))->number();
+        $sin = Degree::of(Number::of(42))->sine();
         $number = $sin->exponential();
 
         $this->assertSame(1.9525390574726629, $number->value());
@@ -138,8 +137,8 @@ class SineTest extends TestCase
 
     public function testBinaryLogarithm()
     {
-        $number = Sine::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->sine()
             ->binaryLogarithm();
 
         $this->assertSame(-0.5796402595724052, $number->value());
@@ -147,8 +146,8 @@ class SineTest extends TestCase
 
     public function testNaturalLogarithm()
     {
-        $number = Sine::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->sine()
             ->naturalLogarithm();
 
         $this->assertSame(-0.4017760116616475, $number->value());
@@ -156,8 +155,8 @@ class SineTest extends TestCase
 
     public function testCommonLogarithm()
     {
-        $number = Sine::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->sine()
             ->commonLogarithm();
 
         $this->assertSame(-0.17448910482575006, $number->value());
@@ -165,8 +164,8 @@ class SineTest extends TestCase
 
     public function testSignum()
     {
-        $number = Sine::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->sine()
             ->signum();
 
         $this->assertSame(1, $number->value());