+++ tests/Geometry/Trigonometry/TangentTest.php
@@ -4,7 +4,6 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Math\Geometry\Trigonometry;
 
 use Innmind\Math\{
-    Geometry\Trigonometry\Tangent,
     Geometry\Angle\Degree,
     Algebra\Number,
 };
@@ -14,7 +13,7 @@ class TangentTest extends TestCase
 {
     public function testInterface()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)));
+        $tan = Degree::of(Number::of(42))->tangent();
 
         $this->assertSame(0.9004040442978399, $tan->number()->value());
         $this->assertSame('tan(42°)', $tan->toString());
@@ -22,7 +21,7 @@ class TangentTest extends TestCase
 
     public function testEquals()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
 
         $this->assertTrue($tan->equals($tan));
         $this->assertTrue($tan->equals(Number::of(0.9004040442978399)));
@@ -31,7 +30,7 @@ class TangentTest extends TestCase
 
     public function testHigherThan()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
 
         $this->assertFalse($tan->higherThan($tan));
         $this->assertFalse($tan->higherThan(Number::of(0.9004040442978399)));
@@ -40,7 +39,7 @@ class TangentTest extends TestCase
 
     public function testAdd()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->add(Number::of(1));
 
         $this->assertSame(1.90040404429784, $number->value());
@@ -48,7 +47,7 @@ class TangentTest extends TestCase
 
     public function testSubtract()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->subtract(Number::of(66));
 
         $this->assertSame(-65.09959595570216, $number->value());
@@ -56,7 +55,7 @@ class TangentTest extends TestCase
 
     public function testDivideBy()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->divideBy(Number::of(2));
 
         $this->assertSame(0.45020202214891997, $number->value());
@@ -64,7 +63,7 @@ class TangentTest extends TestCase
 
     public function testMulitplyBy()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->multiplyBy(Number::of(2));
 
         $this->assertSame(1.8008080885956799, $number->value());
@@ -72,7 +71,7 @@ class TangentTest extends TestCase
 
     public function testRound()
     {
-        $number = Tangent::of(Degree::of(Number::of(42)))->number();
+        $number = Degree::of(Number::of(42))->tangent();
 
         $this->assertEquals(0.9, $number->roundUp(1)->value());
         $this->assertEquals(0.9, $number->roundDown(1)->value());
@@ -82,7 +81,7 @@ class TangentTest extends TestCase
 
     public function testFloor()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->floor();
 
         $this->assertSame(0.0, $number->value());
@@ -90,7 +89,7 @@ class TangentTest extends TestCase
 
     public function testCeil()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->ceil();
 
         $this->assertSame(1.0, $number->value());
@@ -98,7 +97,7 @@ class TangentTest extends TestCase
 
     public function testModulo()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->modulo(Number::of(3));
 
         $this->assertSame(0.9004040442978399, $number->value());
@@ -106,7 +105,7 @@ class TangentTest extends TestCase
 
     public function testAbsolute()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->absolute();
 
         $this->assertSame(0.9004040442978399, $number->value());
@@ -114,7 +113,7 @@ class TangentTest extends TestCase
 
     public function testPower()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->power(Number::of(2));
 
         $this->assertSame(0.8107274429879064, $number->value());
@@ -122,7 +121,7 @@ class TangentTest extends TestCase
 
     public function testSquareRoot()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->squareRoot();
 
         $this->assertSame(0.9488962241983261, $number->value());
@@ -130,7 +129,7 @@ class TangentTest extends TestCase
 
     public function testExponential()
     {
-        $tan = Tangent::of(Degree::of(Number::of(42)))->number();
+        $tan = Degree::of(Number::of(42))->tangent();
         $number = $tan->exponential();
 
         $this->assertSame(2.4605971005633154, $number->value());
@@ -138,8 +137,8 @@ class TangentTest extends TestCase
 
     public function testBinaryLogarithm()
     {
-        $number = Tangent::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->tangent()
             ->binaryLogarithm();
 
         $this->assertSame(-0.1513555580027417, $number->value());
@@ -147,8 +146,8 @@ class TangentTest extends TestCase
 
     public function testNaturalLogarithm()
     {
-        $number = Tangent::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->tangent()
             ->naturalLogarithm();
 
         $this->assertSame(-0.10491167829167766, $number->value());
@@ -156,8 +155,8 @@ class TangentTest extends TestCase
 
     public function testCommonLogarithm()
     {
-        $number = Tangent::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->tangent()
             ->commonLogarithm();
 
         $this->assertEqualsWithDelta(
@@ -169,8 +168,8 @@ class TangentTest extends TestCase
 
     public function testSignum()
     {
-        $number = Tangent::of(Degree::of(Number::of(42)))
-            ->number()
+        $number = Degree::of(Number::of(42))
+            ->tangent()
             ->signum();
 
         $this->assertSame(1, $number->value());